requirements.txt
Flask>=3.0,<4.0
gunicorn>=22.0,<23.0
google-cloud-firestore>=2.16,<3.0
requests>=2.31,<3.0
beautifulsoup4>=4.12,<5.0
python-dotenv>=1.0,<2.0